+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Don't flip flop for multiple instances
 * Fix bug where "retain" message sometimes not show correctly
+* Support for custom protocols via external decoder programs
 
 ## 1.5.0